test: address second review round
